pyproject.toml
[tool.poetry]
name = "gym-env-rlot"
version = "0.1.0"
description = ""
authors = ["Jeroen van Lier <jeroencvlier@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
gymnasium = "^0.28.1"
wandb = "^0.16.3"
pandas = "^2.2.0"
ipywidgets = "^8.1.2"
gputil = "^1.4.0"
scikit-learn = "^1.4.0"
tqdm = "^4.66.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
plotly = "^5.18.0"
ray = {extras = ["rllib"], version = "^2.9.2"}
torch = "^2.2.0"
shap = "^0.44.1"
pytest = "^8.0.2"
icecream = "^2.1.3"
umap-learn = "^0.5.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
